Add logging to testApplyingCouponToCart() function and modify ShoppingCart class to allow access to products array
This commit adds logging to the testApplyingCouponToCart() function to show the contents of the shopping cart before and after a coupon is applied, as well as the total price. It also modifies the ShoppingCart class to allow access to the products array via a new getProducts() method.

The following changes were made:
- Added print statements to testApplyingCouponToCart() function to log shopping cart contents and total price
- Modified ShoppingCart class to add getProducts() method, which returns a copy of the products array